PDF_Summarizer/requirements.txt
langchain-community
langchain
langchain-ollama
streamlit
pypdf
faiss-cpu
tiktoken